+++ src/lib/shim/shim_syscall.c
@@ -1,8 +1,11 @@
+#include "lib/shim/shim_syscall.h"
+
 #include <alloca.h>
 #include <assert.h>
 #include <errno.h>
 #include <stdio.h>
 #include <stdlib.h>
+#include <string.h>
 #include <sys/syscall.h>
 
 #include "lib/logger/logger.h"
@@ -16,10 +19,103 @@
 #include "main/host/syscall/kernel_types.h"
 #include "main/host/syscall_numbers.h"
 
+// Helper macro for calculating immediate offsets in asm.
+#define MCTX_REG_OFFSET(i) (offsetof(mcontext_t, gregs) + sizeof(uint64_t) * (i))
+
 // Never inline, so that the seccomp filter can reliably whitelist a syscall from
 // this function.
 // TODO: Drop if/when we whitelist using /proc/self/maps
-long __attribute__((noinline)) shim_native_syscallv(long n, va_list args) {
+long __attribute__((noinline)) shim_clone(const ucontext_t* ctx, int32_t flags, void* child_stack,
+                                          pid_t* ptid, pid_t* ctid, uint64_t newtls) {
+    if (!child_stack) {
+        panic("clone without a new stack not implemented");
+    }
+    // x86-64 calling conventions require a 16-byte aligned stack
+    if ((uintptr_t)child_stack % 16 != 0) {
+        panic("clone with unaligned new stack not implemented");
+    }
+    // Copy to top of the child stack
+    child_stack -= sizeof(ctx->uc_mcontext);
+    memcpy(child_stack, &ctx->uc_mcontext, sizeof(ctx->uc_mcontext));
+    if ((uintptr_t)child_stack % 16 != 0) {
+        panic("Unimplemented: use padding to realign stack");
+    }
+
+    // Make the clone syscall, and then in the child thread initialize the shim's state,
+    // and then *jump* to the instruction after the original clone syscall instruction.
+    //
+    // Forcing the thread initialization to happen here instead of doing it
+    // lazily ensures that we don't try to install the sigaltstack from
+    // inside the seccomp signal handler. Doing so "works" without error,
+    // but is reverted when the signal handler returns.
+    long rv = 0;
+    register pid_t* r10 __asm__("r10") = ctid;
+    register long r8 __asm__("r8") = newtls;
+    __asm__ __volatile__(
+        "syscall\n"
+        // If in the parent, done with asm.
+        "cmp $0, %%rax\n"
+        "jne shim_native_syscallv_out\n"
+        // Initialize state for this thread
+        "call shim_ensure_init\n"
+        // Restore general purpose registers
+        "mov %c[REG_R8_offset](%%rsp), %%r8\n"
+        "mov %c[REG_R9_offset](%%rsp), %%r9\n"
+        "mov %c[REG_R10_offset](%%rsp), %%r10\n"
+        "mov %c[REG_R11_offset](%%rsp), %%r11\n"
+        "mov %c[REG_R12_offset](%%rsp), %%r12\n"
+        "mov %c[REG_R13_offset](%%rsp), %%r13\n"
+        "mov %c[REG_R14_offset](%%rsp), %%r14\n"
+        "mov %c[REG_R15_offset](%%rsp), %%r15\n"
+        "mov %c[REG_RSI_offset](%%rsp), %%rsi\n"
+        "mov %c[REG_RDI_offset](%%rsp), %%rdi\n"
+        "mov %c[REG_RBX_offset](%%rsp), %%rbx\n"
+        "mov %c[REG_RDX_offset](%%rsp), %%rdx\n"
+        "mov %c[REG_RCX_offset](%%rsp), %%rcx\n"
+        "mov %c[REG_RBP_offset](%%rsp), %%rbp\n"
+        // Save in RAX, to be pushed after
+        // restoring rsp
+        "mov %c[REG_RIP_offset](%%rsp), %%rax\n"
+        // "Free" ctx from the stack
+        "add %[CTX_SIZE], %%rsp\n"
+        // Push original RIP onto the stack.
+        "push %%rax\n"
+        // Not restored:
+        // - RAX: stores the result of the syscall, which we set below.
+        // - Floating point and other special registers: hopefully not needed.
+
+        // Set return value of clone
+        "movq $0, %%rax\n"
+        // Ret to original RIP.
+        "ret\n"
+        "shim_native_syscallv_out:\n"
+        : "=a"(rv)
+        : "a"(SYS_clone), "D"(flags), "S"(child_stack), "d"(ptid), "r"(r10),
+          "r"(r8), [REG_RIP_offset] "i"(MCTX_REG_OFFSET(REG_RIP)),
+          [CTX_SIZE] "i"(sizeof(ctx->uc_mcontext)), [REG_R8_offset] "i"(MCTX_REG_OFFSET(REG_R8)),
+          [REG_R9_offset] "i"(MCTX_REG_OFFSET(REG_R9)),
+          [REG_R10_offset] "i"(MCTX_REG_OFFSET(REG_R10)),
+          [REG_R11_offset] "i"(MCTX_REG_OFFSET(REG_R11)),
+          [REG_R12_offset] "i"(MCTX_REG_OFFSET(REG_R12)),
+          [REG_R13_offset] "i"(MCTX_REG_OFFSET(REG_R13)),
+          [REG_R14_offset] "i"(MCTX_REG_OFFSET(REG_R14)),
+          [REG_R15_offset] "i"(MCTX_REG_OFFSET(REG_R15)),
+          [REG_RSI_offset] "i"(MCTX_REG_OFFSET(REG_RSI)),
+          [REG_RDI_offset] "i"(MCTX_REG_OFFSET(REG_RDI)),
+          [REG_RBX_offset] "i"(MCTX_REG_OFFSET(REG_RBX)),
+          [REG_RDX_offset] "i"(MCTX_REG_OFFSET(REG_RDX)),
+          [REG_RCX_offset] "i"(MCTX_REG_OFFSET(REG_RCX)),
+          [REG_RBP_offset] "i"(MCTX_REG_OFFSET(REG_RBP))
+        : "rcx", "r11", "memory");
+    // Wait for child to initialize itself.
+    shim_newThreadFinish();
+    return rv;
+}
+
+// Never inline, so that the seccomp filter can reliably whitelist a syscall from
+// this function.
+// TODO: Drop if/when we whitelist using /proc/self/maps
+long __attribute__((noinline)) shim_native_syscallv(const ucontext_t* ctx, long n, va_list args) {
     long arg1 = va_arg(args, long);
     long arg2 = va_arg(args, long);
     long arg3 = va_arg(args, long);
@@ -28,86 +124,43 @@ long __attribute__((noinline)) shim_native_syscallv(long n, va_list args) {
     long arg6 = va_arg(args, long);
     long rv;
 
-    // When interposing a clone syscall, we can't return in the new child thread.
-    // Instead we *jump* to just after the original syscall instruction, using
-    // the RIP saved in our SIGSYS signal handler.
-    //
-    // TODO: it'd be cleaner for this to be a separate, dedicated, function.
-    // However right now the actual clone syscall instruction *must* be executed
-    // from this function to pass the seccomp filter.
-    void* clone_rip = NULL;
-    if (n == SYS_clone && (clone_rip = shim_seccomp_take_clone_rip()) != NULL) {
-        // Make the clone syscall, and then in the child thread initialize the shim's state,
-        // and then *jump* to the instruction after the original clone syscall instruction.
-        //
-        // Forcing the thread initialization to happen here instead of doing it
-        // lazily ensures that we don't try to install the sigaltstack from
-        // inside the seccomp signal handler. Doing so "works" without error,
-        // but is reverted when the signal handler returns.
-        //
-        // Note that from the child thread's point of view, many of the general purpose
-        // registers will have different values than they had in the parent thread just-before.
-        // I can't find any documentation on whether the child thread is allowed to make
-        // any assumptions about the state of such registers, but glibc's implementation
-        // of the clone library function doesn't. If we had to, we could save and restore
-        // the other registers in the same way as we are the RIP register.
+    if (n == SYS_clone) {
+        int32_t flags = (int32_t)arg1;
+        void* child_stack = (void*)arg2;
+        pid_t* ptid = (pid_t*)arg3;
+        pid_t* ctid = (pid_t*)arg4;
+        uint64_t newtls = arg5;
+        rv = shim_clone(ctx, flags, child_stack, ptid, ctid, newtls);
+    } else {
+        // r8, r9, and r10 aren't supported as register-constraints in
+        // extended asm templates. We have to use [local register
+        // variables](https://gcc.gnu.org/onlinedocs/gcc/Local-Register-Variables.html)
+        // instead. Calling any functions in between the register assignment and the
+        // asm template could clobber these registers, which is why we don't do the
+        // assignment directly above.
         register long r10 __asm__("r10") = arg4;
         register long r8 __asm__("r8") = arg5;
-        // Store clone_rip in a callee-save register, so that we can call shim_ensure_init
-        // without clobbering it.
-        register long r12 __asm__("r12") = (long)clone_rip;
-        // Store address of shim_ensure_init in a callee-save register.
-        // TODO: We ought to be able to put the literal "shim_ensure_init" in
-        // the inline assembly and have the assembler resolve the address for
-        // us, but on ubuntu 18.04's gcc this ends up generating a relocation it
-        // can't resolve at the link step.
-        register long r13 __asm__("r13") = (long)&shim_ensure_init;
-        __asm__ __volatile__("syscall\n"
-                             // If in the parent, done with asm.
-                             "cmp $0, %%rax\n"
-                             "jne shim_native_syscallv_out\n"
-                             // Initialize state for this thread
-                             "callq *%%r13\n"
-                             // Restore return value of clone
-                             "movq $0, %%rax\n"
-                             // Jump to original clone call site.
-                             "jmp *%%r12\n"
-                             "shim_native_syscallv_out:\n"
+        register long r9 __asm__("r9") = arg6;
+        __asm__ __volatile__("syscall"
                              : "=a"(rv)
-                             : "a"(n), "D"(arg1), "S"(arg2), "d"(arg3), "r"(r10), "r"(r8), "r"(r12), "r"(r13)
+                             : "a"(n), "D"(arg1), "S"(arg2), "d"(arg3), "r"(r10), "r"(r8), "r"(r9)
                              : "rcx", "r11", "memory");
-        // Wait for child to initialize itself.
-        shim_newThreadFinish();
-        return rv;
     }
-
-    // r8, r9, and r10 aren't supported as register-constraints in
-    // extended asm templates. We have to use [local register
-    // variables](https://gcc.gnu.org/onlinedocs/gcc/Local-Register-Variables.html)
-    // instead. Calling any functions in between the register assignment and the
-    // asm template could clobber these registers, which is why we don't do the
-    // assignment directly above.
-    register long r10 __asm__("r10") = arg4;
-    register long r8 __asm__("r8") = arg5;
-    register long r9 __asm__("r9") = arg6;
-    __asm__ __volatile__("syscall"
-                         : "=a"(rv)
-                         : "a"(n), "D"(arg1), "S"(arg2), "d"(arg3), "r"(r10), "r"(r8), "r"(r9)
-                         : "rcx", "r11", "memory");
     return rv;
 }
 
 // Handle to the real syscall function, initialized once at load-time for
 // thread-safety.
-long shim_native_syscall(long n, ...) {
+long shim_native_syscall(const ucontext_t* ctx, long n, ...) {
     va_list args;
     va_start(args, n);
-    long rv = shim_native_syscallv(n, args);
+    long rv = shim_native_syscallv(ctx, n, args);
     va_end(args);
     return rv;
 }
 
-static SysCallReg _shim_emulated_syscall_event(const ShimEvent* syscall_event) {
+static SysCallReg _shim_emulated_syscall_event(const ucontext_t* ctx,
+                                               const ShimEvent* syscall_event) {
 
     struct IPCData* ipc = shim_thisThreadEventIPC();
 
@@ -164,7 +217,7 @@ static SysCallReg _shim_emulated_syscall_event(const ShimEvent* syscall_event) {
                 // Make the original syscall ourselves and use the result.
                 const ShimEventSyscall* syscall = shimevent_getSyscallData(syscall_event);
                 const SysCallReg* regs = syscall->syscall_args.args;
-                SysCallReg rv = {.as_i64 = shim_native_syscall(syscall->syscall_args.number,
+                SysCallReg rv = {.as_i64 = shim_native_syscall(ctx, syscall->syscall_args.number,
                                                                regs[0].as_u64, regs[1].as_u64,
                                                                regs[2].as_u64, regs[3].as_u64,
                                                                regs[4].as_u64, regs[5].as_u64)};
@@ -212,8 +265,8 @@ static SysCallReg _shim_emulated_syscall_event(const ShimEvent* syscall_event) {
                 const ShimEventSyscall* syscall = shimevent_getSyscallData(&res);
                 const SysCallReg* regs = syscall->syscall_args.args;
                 long syscall_rv = shim_native_syscall(
-                    syscall->syscall_args.number, regs[0].as_u64, regs[1].as_u64, regs[2].as_u64,
-                    regs[3].as_u64, regs[4].as_u64, regs[5].as_u64);
+                    ctx, syscall->syscall_args.number, regs[0].as_u64, regs[1].as_u64,
+                    regs[2].as_u64, regs[3].as_u64, regs[4].as_u64, regs[5].as_u64);
                 ShimEvent syscall_complete_event;
                 shimevent_initSysCallComplete(
                     &syscall_complete_event, (SysCallReg){.as_i64 = syscall_rv}, false);
@@ -236,7 +289,7 @@ static SysCallReg _shim_emulated_syscall_event(const ShimEvent* syscall_event) {
     }
 }
 
-long shim_emulated_syscallv(long n, va_list args) {
+long shim_emulated_syscallv(const ucontext_t* ctx, long n, va_list args) {
     bool oldNativeSyscallFlag = shim_swapAllowNativeSyscalls(true);
 
     if (n == SYS_exit) {
@@ -253,22 +306,22 @@ long shim_emulated_syscallv(long n, va_list args) {
     ShimEvent e;
     shimevent_initSyscall(&e, &ev_args);
 
-    SysCallReg retval = _shim_emulated_syscall_event(&e);
+    SysCallReg retval = _shim_emulated_syscall_event(ctx, &e);
 
     shim_swapAllowNativeSyscalls(oldNativeSyscallFlag);
 
     return retval.as_i64;
 }
 
-long shim_emulated_syscall(long n, ...) {
+long shim_emulated_syscall(const ucontext_t* ctx, long n, ...) {
     va_list(args);
     va_start(args, n);
-    long rv = shim_emulated_syscallv(n, args);
+    long rv = shim_emulated_syscallv(ctx, n, args);
     va_end(args);
     return rv;
 }
 
-long shim_syscallv(long n, va_list args) {
+long shim_syscallv(const ucontext_t* ctx, long n, va_list args) {
     shim_ensure_init();
 
     long rv;
@@ -283,22 +336,22 @@ long shim_syscallv(long n, va_list args) {
         trace("Making syscall %ld indirectly; we ask shadow to handle it using the shmem IPC "
               "channel.",
               n);
-        rv = shim_emulated_syscallv(n, args);
+        rv = shim_emulated_syscallv(ctx, n, args);
     } else {
         // The syscall is made directly; ptrace or seccomp will get the syscall signal.
         trace("Making syscall %ld directly; we expect ptrace or seccomp will interpose it, or it "
               "will be handled natively by the kernel.",
               n);
-        rv = shim_native_syscallv(n, args);
+        rv = shim_native_syscallv(ctx, n, args);
     }
 
     return rv;
 }
 
-long shim_syscall(long n, ...) {
+long shim_syscall(const ucontext_t* ctx, long n, ...) {
     va_list(args);
     va_start(args, n);
-    long rv = shim_syscallv(n, args);
+    long rv = shim_syscallv(ctx, n, args);
     va_end(args);
     return rv;
 }